feat: Theme switching

+++ src/commands/theme.rs
@@ -7,6 +7,7 @@ pub fn handle_comamnd(ctx: AppContext, cmd: ThemeCommands) -> Result<()> {
         ThemeCommands::List => list_themes(ctx),
         ThemeCommands::Delete { name } => delete_theme(ctx, name.as_str()),
         ThemeCommands::Current => current_theme(ctx),
+        ThemeCommands::Load { name } => load_theme(ctx, name.as_str()),
         _ => unreachable!(),
     }
 }
@@ -34,7 +35,21 @@ fn create_theme(ctx: AppContext, name: &str) -> Result<()> {
 }
 
 fn delete_theme(ctx: AppContext, name: &str) -> Result<()> {
+    let current_theme = ctx.get_current_theme()?;
+
+    if current_theme == name {
+        println!("[ WARN ] Cannot delete current theme, please switch to another theme to delete '{name}'");
+        return Ok(())
+    }
+
     ctx.delete_theme(name)?;
-    println!("[ \x1b[91mOK\x1b[0m ] '{name}' theme deleted");
+    println!("[ \x1b[92mOK\x1b[0m ] '{name}' theme deleted");
+    Ok(())
+}
+
+fn load_theme(ctx: AppContext, name: &str) -> Result<()> {
+    ctx.switch_current(name)?;
+    println!("[ \x1b[92mOK\x1b[0m ] Switched theme to '{name}'");
+    ctx.create_misisng_theme_files()?;
     Ok(())
 }
